Add custom exceptions for domain cases
