+++ backend/src/main/java/com/servicedesk/lite/error/OrgExceptionHandler.java
@@ -1,0 +1,7 @@
+package com.servicedesk.lite.error;
+
+//@RestControllerAdvice(basePackages = "com.servicedesk.lite.org")
+public class OrgExceptionHandler {
+
+
+}
