+++ backend/src/main/java/com/servicedesk/lite/tickets/exceptions/UnauthorizedCommentAuthorException.java
@@ -1,0 +1,7 @@
+package com.servicedesk.lite.tickets.exceptions;
+
+public class UnauthorizedCommentAuthorException extends RuntimeException {
+    public UnauthorizedCommentAuthorException(String message) {
+        super(message);
+    }
+}
